driver: correctly set 'BYTE_SWAPPING' property on bswap library.
Probably a copy-pasta from when this was migrated from rosbuild.

The 'set_target_properties()' call set the property on the wrong target
(namely: the non-swapping one). This probably did not affect anything, as
byte swaps are performed at the simple-message level, and that target
(configured earlier in the build file) was correctly configured with the
BYTE_SWAPPING property.